+++ app/src/main/java/com/bmprj/cointicker/model/Low24h.kt
@@ -8,5 +8,5 @@ import kotlinx.serialization.Serializable
 data class Low24h(
 
     @SerializedName("usd")
-    val usd: Int
+    val usd: Double
 )
