+++ app/src/main/java/com/bmprj/cointicker/view/coin/CoinDetailFragment.kt
@@ -109,11 +109,11 @@ class CoinDetailFragment : BaseFragment<FragmentCoinDetailBinding>(R.layout.frag
                 binding.coinsymbol.text=it.symbol
                 binding.imageView.loadFromUrl(it.image.large)
                 binding.lastUpdate.text=it.lastUpdated
-                binding.priceChange24h.text=it.marketData.currentPrice.usd.toDouble().toString() + "$"
+                binding.priceChange24h.text=it.marketData.currentPrice.usd.toString() + "$"
                 binding.arrow.setArrow(it.marketData.priceChangePercentage24h)
                 binding.precentage24hText.text= String.format("%.1f",it.marketData.priceChangePercentage24h.toFloat())+"%"
-                binding.lowest.text=it.marketData.low24h.usd.toDouble().toString()
-                binding.highest.text=it.marketData.high24h.usd.toDouble().toString()
+                binding.lowest.text=it.marketData.low24h.usd.toString()
+                binding.highest.text=it.marketData.high24h.usd.toString()
                 binding.description.text=it.description.en
             }
